refactor: modularize script execution business logic
