+++ packages/hoppscotch-js-sandbox/src/web/pre-request/index.ts
@@ -1,7 +1,6 @@
 import { FaradayCage } from "faraday-cage"
 import { ConsoleEntry } from "faraday-cage/modules"
 import * as E from "fp-ts/Either"
-import * as TE from "fp-ts/lib/TaskEither"
 import { cloneDeep } from "lodash"
 import { SandboxPreRequestResult, TestResult } from "~/types"
 
@@ -9,30 +8,33 @@ import { defaultModules, pwPreRequestModule } from "~/cage-modules"
 
 import Worker from "./worker?worker&inline"
 
-export const runPreRequestScript = async (
+const runPreRequestScriptWithWebWorker = (
   preRequestScript: string,
-  envs: TestResult["envs"],
-  experimentalScriptingSandbox = true,
+  envs: TestResult["envs"]
+): Promise<E.Either<string, SandboxPreRequestResult>> => {
+  return new Promise((resolve) => {
+    const worker = new Worker()
+
+    // Listen for the results from the web worker
+    worker.addEventListener("message", (event: MessageEvent) => {
+      worker.terminate()
+      return resolve(event.data.results)
+    })
+
+    // Send the script to the web worker
+    worker.postMessage({
+      preRequestScript,
+      envs,
+    })
+  })
+}
+
+const runPreRequestScriptWithFaradayCage = async (
+  preRequestScript: string,
+  envs: TestResult["envs"]
 ): Promise<E.Either<string, SandboxPreRequestResult>> => {
   const consoleEntries: ConsoleEntry[] = []
   let finalEnvs = envs
-
-  if (!experimentalScriptingSandbox) {
-    return new Promise((resolve) => {
-      const worker = new Worker()
-
-      // Listen for the results from the web worker
-      worker.addEventListener("message", (event: MessageEvent) =>
-        resolve(event.data.results),
-      )
-
-      // Send the script to the web worker
-      worker.postMessage({
-        preRequestScript,
-        envs,
-      })
-    })
-  }
 
   const cage = await FaradayCage.create()
 
@@ -40,6 +42,7 @@ export const runPreRequestScript = async (
     ...defaultModules({
       handleConsoleEntry: (consoleEntry) => consoleEntries.push(consoleEntry),
     }),
+
     pwPreRequestModule({
       envs: cloneDeep(envs),
       handleSandboxResults: ({ envs }) => (finalEnvs = envs),
@@ -52,14 +55,23 @@ export const runPreRequestScript = async (
       typeof result.err === "object" &&
       "message" in result.err
     ) {
-      return TE.left(`Script execution failed: ${result.err.message}`)()
+      return E.left(`Script execution failed: ${result.err.message}`)
     }
 
-    return TE.left(`Script execution failed: ${String(result.err)}`)()
+    return E.left(`Script execution failed: ${String(result.err)}`)
   }
 
-  return TE.right({
+  return E.right({
     envs: finalEnvs,
     consoleEntries,
-  })()
+  })
 }
+
+export const runPreRequestScript = async (
+  preRequestScript: string,
+  envs: TestResult["envs"],
+  experimentalScriptingSandbox = true
+): Promise<E.Either<string, SandboxPreRequestResult>> =>
+  experimentalScriptingSandbox
+    ? runPreRequestScriptWithFaradayCage(preRequestScript, envs)
+    : runPreRequestScriptWithWebWorker(preRequestScript, envs)
