+++ packages/hoppscotch-js-sandbox/src/web/test-runner/index.ts
@@ -1,5 +1,10 @@
+import { FaradayCage } from "faraday-cage"
+import { ConsoleEntry } from "faraday-cage/modules"
 import * as E from "fp-ts/Either"
+import { cloneDeep } from "lodash-es"
 
+import { defaultModules, pwPostRequestModule } from "~/cage-modules"
+import { preventCyclicObjects } from "~/shared-utils"
 import {
   SandboxTestResult,
   TestDescriptor,
@@ -7,21 +12,34 @@ import {
   TestResult,
 } from "~/types"
 
-import { FaradayCage } from "faraday-cage"
-import { ConsoleEntry } from "faraday-cage/modules"
-import * as TE from "fp-ts/lib/TaskEither"
-import { cloneDeep } from "lodash-es"
-
-import { defaultModules, pwPostRequestModule } from "~/cage-modules"
-import { preventCyclicObjects } from "~/shared-utils"
-
 import Worker from "./worker?worker&inline"
 
-export const runTestScript = async (
+const runTestScriptWithWebWorker = (
   testScript: string,
   envs: TestResult["envs"],
-  response: TestResponse,
-  experimentalScriptingSandbox = true,
+  response: TestResponse
+): Promise<E.Either<string, SandboxTestResult>> => {
+  return new Promise((resolve) => {
+    const worker = new Worker()
+
+    // Listen for the results from the web worker
+    worker.addEventListener("message", (event: MessageEvent) =>
+      resolve(event.data.results)
+    )
+
+    // Send the script to the web worker
+    worker.postMessage({
+      testScript,
+      envs,
+      response,
+    })
+  })
+}
+
+const runTestScriptWithFaradayCage = async (
+  testScript: string,
+  envs: TestResult["envs"],
+  response: TestResponse
 ): Promise<E.Either<string, SandboxTestResult>> => {
   const testRunStack: TestDescriptor[] = [
     { descriptor: "root", expectResults: [], children: [] },
@@ -30,30 +48,6 @@ export const runTestScript = async (
   let finalEnvs = envs
   let finalTestResults = testRunStack
   const consoleEntries: ConsoleEntry[] = []
-
-  const responseObjHandle = preventCyclicObjects(response)
-
-  if (E.isLeft(responseObjHandle)) {
-    return TE.left(`Response marshalling failed: ${responseObjHandle.left}`)()
-  }
-
-  if (!experimentalScriptingSandbox) {
-    return new Promise((resolve) => {
-      const worker = new Worker()
-
-      // Listen for the results from the web worker
-      worker.addEventListener("message", (event: MessageEvent) =>
-        resolve(event.data.results),
-      )
-
-      // Send the script to the web worker
-      worker.postMessage({
-        testScript,
-        envs,
-        response,
-      })
-    })
-  }
 
   const cage = await FaradayCage.create()
 
@@ -65,7 +59,7 @@ export const runTestScript = async (
     pwPostRequestModule({
       envs: cloneDeep(envs),
       testRunStack: cloneDeep(testRunStack),
-      response: responseObjHandle.right as TestResponse,
+      response,
       handleSandboxResults: ({ envs, testRunStack }) => {
         finalEnvs = envs
         finalTestResults = testRunStack
@@ -73,25 +67,38 @@ export const runTestScript = async (
     }),
   ])
 
-  console.error(`Done with tests`)
-
   if (result.type === "error") {
-    if (result.type === "error") {
-      if (
-        result.err !== null &&
-        typeof result.err === "object" &&
-        "message" in result.err
-      ) {
-        return TE.left(`Script execution failed: ${result.err.message}`)()
-      }
-
-      return TE.left(`Script execution failed: ${String(result.err)}`)()
+    if (
+      result.err !== null &&
+      typeof result.err === "object" &&
+      "message" in result.err
+    ) {
+      return E.left(`Script execution failed: ${result.err.message}`)
     }
+
+    return E.left(`Script execution failed: ${String(result.err)}`)
   }
 
-  return TE.right(<SandboxTestResult>{
+  return E.right(<SandboxTestResult>{
     tests: finalTestResults[0],
     envs: finalEnvs,
     consoleEntries,
-  })()
+  })
+}
+
+export const runTestScript = async (
+  testScript: string,
+  envs: TestResult["envs"],
+  response: TestResponse,
+  experimentalScriptingSandbox = true
+): Promise<E.Either<string, SandboxTestResult>> => {
+  const responseObjHandle = preventCyclicObjects<TestResponse>(response)
+
+  if (E.isLeft(responseObjHandle)) {
+    return E.left(`Response marshalling failed: ${responseObjHandle.left}`)
+  }
+
+  return experimentalScriptingSandbox
+    ? runTestScriptWithFaradayCage(testScript, envs, responseObjHandle.right)
+    : runTestScriptWithWebWorker(testScript, envs, responseObjHandle.right)
 }
